1. 优化reduceRatio, 目前使用1.2比较好，太大的话会导致特征匹配误差比较大；2. 检测匹配后的位置，来判断匹配是否正确，该方法不一定是最好的，目前可以过滤掉一些误判的结果; 3.  按照queryImage的截图顺序来检测，所以要保证queryImage中的截图顺序与测试的顺序是一致的

+++ testGameByOpencv.py
@@ -10,21 +10,29 @@ import time
 import subprocess
 import glob
 import math
+from os.path import getmtime
 import traceback
 from libs.AdbCommand import screencap
 from libs.GetDeviceInfo import getDeviceInfo
 from PIL import Image
 
 currPath = os.getcwd()
-queryImageRoot = os.path.join(currPath, 'queryImage')
-#x_reduceRatio = 2   #截图横坐标缩放倍数
-#y_reduceRatio = 2   #截图纵坐标缩放倍数
-thumbnailSize = (540.0, 960.0)
+queryImageRoot = os.path.join(currPath, 'imageFile/queryImage')
+matchImageRoot = os.path.join(currPath, 'imageFile/matchImage')
+sceneImageRoot = os.path.join(currPath, 'imageFile/sceneImage')
+#thumbnail size = 1.5
+#portrait_thumbnailSize = (720.0, 1280.0)
+#landscape_thumbnailSize = (1280.0, 720.0)
+#thumbnail size = 1.2
+portrait_thumbnailSize = (900.0, 1600.0)
+landscape_thumbnailSize = (1600.0, 900.0)
+MIN_MATCH_COUNT = 4
 finishFlag = False
 #deviceName = '5LM7N16224000261'
 #deviceName = 'KWG5T17105003967'  #hw P9
 #deviceName = '63a9bca7'  #vivo
-deviceName = 'LGH8689e43a709'  #LG
+#deviceName = 'LGH8689e43a709'  #LG
+deviceName = '635f9505'    #MI 5
 
 
 def signal_handler(signal, frame):
@@ -68,11 +76,12 @@ def getImgCordinate(filePath, sceneFilePath, flag):
     kp2, desc2 = detector.detectAndCompute(img2, None)
     raw_matches = matcher.knnMatch(desc1, trainDescriptors=desc2, k=2) #2特征之匹配
     p1, p2, kp_pairs = filter_matches(kp1, kp2, raw_matches)
+    print '###p1 len is %d, p2 len is %d, kp_pairs len is %d' % (len(p1), len(p2), len(kp_pairs))
 
-    if len(p1) >= 4:
+    if len(p1) >= MIN_MATCH_COUNT:
         H, status = cv2.findHomography(p1, p2, cv2.RANSAC, 5.0)  #获取转换矩阵
         print '****%d / %d inliers/matched' % (np.sum(status), len(status))
-        if np.sum(status) <= len(status)/2:
+        if np.sum(status) < len(status)/2:
             return None, None
     else:
         H, status = None, None
@@ -85,18 +94,25 @@ def getImgCordinate(filePath, sceneFilePath, flag):
     obj_corners = obj_corners.reshape(1, -1, 2)
     scene_corners = cv2.perspectiveTransform(obj_corners, H)  #坐标映射
     scene_corners = scene_corners.reshape(-1, 2)
+    print int(round(scene_corners[3][0])), int(round(scene_corners[3][1])), int(round(scene_corners[1][0])), int(round(scene_corners[1][1]))
+    rectangle_width = int(round(scene_corners[1][0])) - int(round(scene_corners[3][0]))
+    rectangle_height = int(round(scene_corners[1][1])) - int(round(scene_corners[3][1]))
+
     img3 = cv2.rectangle(img2, (int(round(scene_corners[3][0])), int(round(scene_corners[3][1]))), (int(round(scene_corners[1][0])), int(round(scene_corners[1][1]))), (0, 255, 0), 3)
-    resultFilePath = os.path.join(queryImageRoot, pkName, flag+'_match.png')
+    resultFilePath = os.path.join(matchImageRoot, pkName, flag+'_match.png')
     cv2.imwrite(resultFilePath, img3)   #保存在原始截图上标记query pic位置的图片
+    if (abs(rectangle_width) < w1/3 and abs(rectangle_height) < h1/2) or (rectangle_width > w1*1.5 and rectangle_height > h1*1.5):
+        print 'rectangle_width is %d, rectangle_height is %d' % (rectangle_width, rectangle_height)
+        return None, None
     mid_cordinate_x = int(round((scene_corners[3][0]+scene_corners[1][0])/2))   #计算中心坐标
     mid_cordinate_y = int(round((scene_corners[3][1]+scene_corners[1][1])/2))
     #通过特征提取的图片被缩放了，所以计算真正比例的坐标需要在将计算得到的中心坐标在放大reduceRatio
     real_x = math.floor(mid_cordinate_x * x_reduceRatio)
     real_y = math.floor(mid_cordinate_y * y_reduceRatio)
-    print 'real x: %d, real y: %d' % (real_x, real_y)
+    #print 'real x: %d, real y: %d' % (real_x, real_y)
     return real_x, real_y
 
-def thumbnail_pic(path):
+def thumbnail_pic(path, thumbnailSize):
     savePath = None
     try:
         im = Image.open(path)
@@ -117,9 +133,36 @@ def get_resolution():
     max_height = int(splitInfo[1].split()[0])
     return (max_width, max_height)
 
+#小米5S检测横竖屏的时候，返回两行
+#SurfaceOrientation: 0, 第一行结果一直是0
+#SurfaceOrientation: 0/1, 第二行，横屏为1，竖屏为0
+def check_portrait_landscape():
+    checkCmd = "adb -s %s shell dumpsys input | grep 'SurfaceOrientation'  | awk '{ print $2 }'" % deviceName
+    result = os.popen(checkCmd , 'r')
+    content = result.readlines()
+    result.close()
+
+    screenType = 'portrait'  #默认竖屏
+    print 'check portrait result is: %s' % repr(content)
+    if len(content) > 0:
+        orientationValue = content[len(content)-1].replace('\r\n', '')
+        screenTypeValue = int(orientationValue)
+        if screenTypeValue == 0 or screenTypeValue == 2:
+            screenType = 'portrait'   #竖屏
+        elif screenTypeValue == 1 or screenTypeValue == 3:
+            screenType = 'landscape'  #横屏
+        else:
+            print 'dumpsys input has problem.......'
+    return screenType
+
+#360用户登录
+#用户名：adb shell input text 568042716@qq.com
+#回车： adb shell keyevent 66
+#密码： adb shell input text 123658
+
 if __name__ == '__main__':
-    x_reduceRatio = 2   #截图横坐标缩放倍数
-    y_reduceRatio = 2   #截图纵坐标缩放倍数
+    x_reduceRatio = 1   #截图横坐标缩放倍数
+    y_reduceRatio = 1   #截图纵坐标缩放倍数
     queryPkDic = {}
     queryPkList = []
     for packagePath in glob.glob(os.path.join(queryImageRoot, '*/')):
@@ -128,7 +171,7 @@ if __name__ == '__main__':
     for packagePath in queryPkList:
         packageName = packagePath[packagePath.rfind('/')+1:]
         queryPkDic[packageName] = []
-        for queryPic in glob.glob(os.path.join(packagePath, '*-thumbnail.png')):
+        for queryPic in sorted(glob.glob(os.path.join(packagePath, '*-thumbnail.png')), key=getmtime):
             queryPkDic[packageName].append(queryPic)
 
     clearLogCmd = 'adb shell logcat -c'
@@ -138,17 +181,15 @@ if __name__ == '__main__':
     proc = subprocess.Popen(readLogcatCmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     pkName = 'noName'
     resolution = get_resolution()
-    x_reduceRatio = round(resolution[0]/thumbnailSize[0],2)
-    y_reduceRatio = round(resolution[1]/thumbnailSize[1],2)
-    print 'resolution is %s, x_reduceRatio is %s, y_reduceRatio is %s' % (str(resolution), str(x_reduceRatio), str(y_reduceRatio))
 
     for line in proc.stdout:
-        if 'PluginAutoTest' in line and 'install success' in line:
+        if 'ActivityManager' in line and 'Displayed' in line:
             line = line.replace('\r\n', '')
             splitLine = line.split(': ')
             pkItem = splitLine[1]
             pkSplit = pkItem.split()
-            pkName = pkSplit[0]
+            startInfo = pkSplit[1]
+            pkName = startInfo[0:startInfo.find('/')]
             print pkName
             break
     proc.kill()
@@ -157,46 +198,70 @@ if __name__ == '__main__':
 
 
     queryPkImageRoot = os.path.join(queryImageRoot, pkName)
+    matchPkImageRoot = os.path.join(matchImageRoot, pkName)
+    scenePkImageRoot = os.path.join(sceneImageRoot, pkName)
     if os.path.isdir(queryPkImageRoot) is False:
         os.makedirs(queryPkImageRoot)
+    if os.path.isdir(matchPkImageRoot) is False:
+        os.makedirs(matchPkImageRoot)
+    else:
+        os.system('rm -rf %s/*.png' % matchPkImageRoot)
+    if os.path.isdir(scenePkImageRoot) is False:
+        os.makedirs(scenePkImageRoot)
+    else:
+        os.system('rm -rf %s/*.png' % scenePkImageRoot)
 
     startTime = time.time()
     if queryPkDic.has_key(pkName):
         signal.signal(signal.SIGINT, signal_handler)
         picNo = 0
         time.sleep(8)
-        while finishFlag is False:
-            hasClick = False
-            sceneFilePath = os.path.join(queryImageRoot, pkName, 'screen-%d.png' % picNo)
-            screencap(sceneFilePath, None)
+
+        screenType = check_portrait_landscape()
+        if cmp(screenType, 'landscape') == 0:
+            thumbnailSize = landscape_thumbnailSize
+        else:
+            thumbnailSize = portrait_thumbnailSize
+        print 'thumbnailSize is %s' % str(thumbnailSize)
+        if cmp(screenType, 'landscape') == 0:
+            x_reduceRatio = round(resolution[0]/thumbnailSize[1],2)
+            y_reduceRatio = round(resolution[1]/thumbnailSize[0],2)
+        else:
+            x_reduceRatio = round(resolution[0]/thumbnailSize[0],2)
+            y_reduceRatio = round(resolution[1]/thumbnailSize[1],2)
+        print 'resolution is %s, x_reduceRatio is %s, y_reduceRatio is %s' % (str(resolution), str(x_reduceRatio), str(y_reduceRatio))
+
+
+        for queryImageThumbnailPath in queryPkDic[pkName]:
+            print '#####now query image is %s' % queryImageThumbnailPath
             if finishFlag is True:
                 break
-            sceneFileThumbnailPath = thumbnail_pic(sceneFilePath)
-            if sceneFileThumbnailPath is None:
-                break
-            os.remove(sceneFilePath)    #删除未缩放的截图
-            for queryImageThumbnailPath in queryPkDic[pkName]:
-                if finishFlag is True:
-                    break
-                (x,y) = getImgCordinate(queryImageThumbnailPath, sceneFileThumbnailPath, 'screen-%d' % picNo)
-                if (x,y) != (None, None):
-                    print 'matching %s' % queryImageThumbnailPath
-                    print 'click %d, %d' % (x, y)
-                    os.system('adb shell input tap %d %d' % (x, y))
-                    hasClick = True
-                    queryPkDic[pkName].remove(queryImageThumbnailPath)
-                    break
-            #os.remove(sceneFileThumbnailPath)    #删除缩放后的截图，只保留match后的缩放截图
-            if len(queryPkDic[pkName]) == 0:
-                print 'finish query %s all pictures' % pkName
-                break
-            if hasClick == False:
-                #删除匹配失败的缩略截图
-                #os.remove(sceneFileThumbnailPath)
-                time.sleep(5)
-            else:
+            if 'goBack-thumbnail.png' in queryImageThumbnailPath:
+                print 'click goBack'
+                os.system('adb shell input keyevent 4')
                 time.sleep(2)
-            picNo += 1
+            else:
+                maxCmpCount = 20
+                while maxCmpCount > 0 and finishFlag is False:
+                    sceneFilePath = os.path.join(scenePkImageRoot, 'screen-%d.png' % picNo)
+                    screencap(sceneFilePath, None)
+                    sceneFileThumbnailPath = thumbnail_pic(sceneFilePath, thumbnailSize)
+                    if sceneFileThumbnailPath is None:
+                        maxCmpCount -= 1
+                        continue
+                    os.remove(sceneFilePath)    #删除未缩放的截图
+                    (x,y) = getImgCordinate(queryImageThumbnailPath, sceneFileThumbnailPath, 'screen-%d' % picNo)
+                    picNo += 1
+                    if (x,y) != (None, None):
+                        print 'matching %s' % queryImageThumbnailPath
+                        print 'click %d, %d' % (x, y)
+                        os.system('adb shell input tap %d %d' % (x, y))
+                        time.sleep(2)
+                        break
+                    else:
+                        maxCmpCount -= 1
+                        time.sleep(2)
+                    #os.remove(sceneFileThumbnailPath)    #删除缩放后的截图，只保留match后的缩放截图
 
     endTime = time.time()
     print 'spend time is %s' % str(round(endTime-startTime, 3))
